fix: resolve remaining build errors in autonomous features
- advanced_runtime.go: use EnvInspector.Inspect() not Snapshot()
- advanced_runtime.go: PersistencePlanner.Plan takes (string, int) not map
- advanced_runtime.go: remove MemoryExecutor.Execute (no such method)
- advanced_runtime.go: OnAdapt callback takes EnvProfile not map
- runtime.go: use TriggerPropagation not Propagate
- runtime.go: remove non-existent RecordActivity call
- runtime.go: remove non-existent SinkResistance.Evaluate call
- enhanced_stubs.go: cast NumGoroutine() int to float64

+++ pkg/agent/enhanced_stubs.go
@@ -193,7 +193,7 @@ func (pm *PerformanceMonitor) GetLatestMetrics() *PerformanceMetrics {
 	runtime.ReadMemStats(&m)
 	return &PerformanceMetrics{
 		Timestamp:   time.Now(),
-		CPUUsage:    runtime.NumGoroutine(),
+		CPUUsage:    float64(runtime.NumGoroutine()),
 		MemoryUsage: m.Alloc,
 	}
 }
